+++ src/Components/AllContatcsTable/AllContactsTable.tsx
@@ -8,7 +8,7 @@ import {
     MAX_TAG_LIST_LENGTH_BEFORE_SCROLLABLE,
     TAG_LIST_HEIGHT,
     TAG_ROW_HEIGHT,
-} from "../../helpers/constants";
+} from "../../Constants/heights";
 import { getTotalItemSize } from "../TagList/TagList";
 import classNames from "classnames/bind";
